Introduce CPLEX dependency
Separate interface and cplex-dependent code
