+++ lib/src/internal/stub_cfrs_cplex_inl.hpp
@@ -1,0 +1,12 @@
+#include "cluster_first_route_second.h"
+
+namespace vrp {
+namespace detail {
+
+/// Cluster-first Route-second stub
+std::vector<Solution> cfrs_impl(const Problem& prob, size_t count) {
+    return {};
+}
+
+}  // detail
+}  // vrp
